the whole project is finished, only the footer is missing

+++ src/app/albums/components/list-photo/list-photo.component.ts
@@ -2,6 +2,10 @@ import { Component, OnInit, Input } from '@angular/core';
 import { Photo } from '../../models/photo.model';
 import { AlbumService } from '../../services/album.service';
 import { ModalService } from '../../services/modal.service';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogAceptCancelComponent } from '../dialog-acept-cancel/dialog-acept-cancel.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { SnackBarComponent } from '../snack-bar/snack-bar.component';
 
 @Component({
   selector: 'app-list-photo',
@@ -15,8 +19,19 @@ export class ListPhotoComponent implements OnInit {
 
   constructor(
     private modalSevice: ModalService,
-    private albumService: AlbumService
+    private albumService: AlbumService,
+    private _snackBar: MatSnackBar,
+    public dialog: MatDialog
   ) {}
+
+  openDialog() {
+    const dialogRef = this.dialog.open(DialogAceptCancelComponent);
+    return dialogRef.afterClosed();
+  }
+
+  openSnackBar() {
+    this._snackBar.openFromComponent(SnackBarComponent, { duration: 2500 });
+  }
 
   handleSelectPhoto(selectedPhoto: Photo): void {
     const body = document.querySelector('body');
@@ -27,18 +42,23 @@ export class ListPhotoComponent implements OnInit {
   }
 
   handleDeletePhoto(deletedPhoto: Photo): void {
-    this.albumService.deletePhoto(deletedPhoto).subscribe({
-      next: (_: Photo) => {
-        let photos = localStorage.getItem('deletedPhotos');
-        if (!photos)
-          localStorage.setItem('deletedPhotos', deletedPhoto.id.toString());
-        else {
-          let photosArray = photos.split(',');
-          photosArray.push(deletedPhoto.id.toString());
-          localStorage.setItem('deletedPhotos', photosArray.toString());
-        }
-        this.albumService.getPhotosByAlbumId(deletedPhoto.albumId);
-      },
+    this.openDialog().subscribe((result) => {
+      if (result === true) {
+        this.albumService.deletePhoto(deletedPhoto).subscribe({
+          next: (_: Photo) => {
+            let photos = localStorage.getItem('deletedPhotos');
+            if (!photos)
+              localStorage.setItem('deletedPhotos', deletedPhoto.id.toString());
+            else {
+              let photosArray = photos.split(',');
+              photosArray.push(deletedPhoto.id.toString());
+              localStorage.setItem('deletedPhotos', photosArray.toString());
+            }
+            this.openSnackBar();
+            this.albumService.getPhotosByAlbumId(deletedPhoto.albumId);
+          },
+        });
+      }
     });
   }
